move @waku/message-hash into @waku/core
