+++ packages/core/src/index.ts
@@ -28,3 +28,5 @@ export { KeepAliveManager } from "./lib/keep_alive_manager.js";
 export { StreamManager } from "./lib/stream_manager.js";
 
 export { MetadataCodec, wakuMetadata } from "./lib/metadata/index.js";
+
+export * from "./lib/message_hash.js";
