+++ packages/core/src/lib/message_hash.ts
