--- packages/message-hash/.mocha.reporters.json
@@ -1,6 +1,0 @@
-{
-  "reporterEnabled": "spec, allure-mocha",
-  "allureMochaReporter": {
-    "outputDir": "allure-results"
-  }
-}
