--- packages/message-hash/.prettierignore
@@ -1,5 +1,0 @@
-build
-bundle
-dist
-node_modules
-CHANGELOG.md
