--- packages/message-hash/CHANGELOG.md
@@ -1,122 +1,0 @@
-# Changelog
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from 0.0.6 to 0.0.7
-  * devDependencies
-    * @waku/interfaces bumped from 0.0.13 to 0.0.14
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from 0.0.7 to 0.0.8
-  * devDependencies
-    * @waku/interfaces bumped from 0.0.14 to 0.0.15
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from 0.0.8 to 0.0.9
-  * devDependencies
-    * @waku/interfaces bumped from 0.0.15 to 0.0.16
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from 0.0.9 to 0.0.10
-  * devDependencies
-    * @waku/interfaces bumped from 0.0.16 to 0.0.17
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from 0.0.12 to 0.0.13
-  * devDependencies
-    * @waku/interfaces bumped from 0.0.19 to 0.0.20
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from 0.0.13 to 0.0.14
-  * devDependencies
-    * @waku/interfaces bumped from 0.0.20 to 0.0.21
-
-## [0.1.8](https://github.com/waku-org/js-waku/compare/message-hash-v0.1.7...message-hash-v0.1.8) (2023-10-16)
-
-
-### Features
-
-* Add Firefox and Webkit to karma  ([#1598](https://github.com/waku-org/js-waku/issues/1598)) ([d9e4bcb](https://github.com/waku-org/js-waku/commit/d9e4bcbe3f7bcc092f20621bd362d76426701dab))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from 0.0.11 to 0.0.12
-  * devDependencies
-    * @waku/interfaces bumped from 0.0.18 to 0.0.19
-
-## [0.1.7](https://github.com/waku-org/js-waku/compare/message-hash-v0.1.6...message-hash-v0.1.7) (2023-09-11)
-
-
-### ⚠ BREAKING CHANGES
-
-* @waku/relay ([#1316](https://github.com/waku-org/js-waku/issues/1316))
-
-### Features
-
-* @waku/relay ([#1316](https://github.com/waku-org/js-waku/issues/1316)) ([50c2c25](https://github.com/waku-org/js-waku/commit/50c2c2540f3c5ff78d93f3fea646da0eee246e17))
-* Implement deterministic message hash logic ([fe57461](https://github.com/waku-org/js-waku/commit/fe574613290875d1c142aa51c18641346208401d))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from 0.0.10 to 0.0.11
-  * devDependencies
-    * @waku/interfaces bumped from 0.0.17 to 0.0.18
-
-## [0.1.2](https://github.com/waku-org/js-waku/compare/message-hash-v0.1.1...message-hash-v0.1.2) (2023-05-18)
-
-
-### ⚠ BREAKING CHANGES
-
-* @waku/relay ([#1316](https://github.com/waku-org/js-waku/issues/1316))
-
-### Features
-
-* @waku/relay ([#1316](https://github.com/waku-org/js-waku/issues/1316)) ([50c2c25](https://github.com/waku-org/js-waku/commit/50c2c2540f3c5ff78d93f3fea646da0eee246e17))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from * to 0.0.6
-  * devDependencies
-    * @waku/interfaces bumped from * to 0.0.13
-
-## [0.1.1](https://github.com/waku-org/js-waku/compare/message-hash-v0.1.0...message-hash-v0.1.1) (2023-05-09)
-
-
-### Features
-
-* Implement deterministic message hash logic ([fe57461](https://github.com/waku-org/js-waku/commit/fe574613290875d1c142aa51c18641346208401d))
-
-
-### Dependencies
-
-* The following workspace dependencies were updated
-  * dependencies
-    * @waku/utils bumped from * to 0.0.5
-  * devDependencies
-    * @waku/interfaces bumped from * to 0.0.12
