--- packages/message-hash/README.md
@@ -1,26 +1,0 @@
-[![NPM](https://nodei.co/npm/@waku/message-hash.png)](https://npmjs.org/package/@waku/message-hash)
-
-![GitHub Action](https://img.shields.io/github/workflow/status/waku-org/js-waku/CI)
-[![Discord chat](https://img.shields.io/discord/864066763682218004.svg?logo=discord&colorB=7289DA)](https://discord.gg/Nrac59MfSX)
-
-# @waku/message-hash
-
-TypeScript implementation of the _Deterministic Message Hashing_ as specified in [14/WAKU2-MESSAGE](https://rfc.vac.dev/spec/14/).
-
-See [JS-Waku README](https://github.com/waku-org/js-waku) for more information.
-
-## Contributing
-
-See [CONTRIBUTING.md](https://github.com/waku-org/js-waku/blob/master/CONTRIBUTING.md).
-
-## License
-
-Licensed and distributed under either of
-
-- MIT license: [LICENSE-MIT](https://github.com/waku-org/js-waku/blob/master/LICENSE-MIT) or http://opensource.org/licenses/MIT
-
-or
-
-- Apache License, Version 2.0, ([LICENSE-APACHE-v2](https://github.com/waku-org/js-waku/blob/master/LICENSE-APACHE-v2) or http://www.apache.org/licenses/LICENSE-2.0)
-
-at your option. These files may not be copied, modified, or distributed except according to those terms.
