--- packages/message-hash/package.json
@@ -1,86 +1,0 @@
-{
-  "name": "@waku/message-hash",
-  "version": "0.1.10",
-  "description": "TypeScript implementation of the Deterministic Message Hashing as specified in 14/WAKU2-MESSAGE",
-  "types": "./dist/index.d.ts",
-  "module": "./dist/index.js",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "type": "module",
-  "homepage": "https://github.com/waku-org/js-waku/tree/master/packages/message-hash#readme",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/waku-org/js-waku.git"
-  },
-  "bugs": {
-    "url": "https://github.com/waku-org/js-waku/issues"
-  },
-  "license": "MIT OR Apache-2.0",
-  "keywords": [
-    "waku",
-    "decentralised",
-    "communication",
-    "web3",
-    "ethereum",
-    "dapps"
-  ],
-  "scripts": {
-    "build": "run-s build:**",
-    "build:esm": "tsc",
-    "build:bundle": "rollup --config rollup.config.js",
-    "fix": "run-s fix:*",
-    "fix:lint": "eslint src *.js --fix",
-    "check": "run-s check:*",
-    "check:tsc": "tsc -p tsconfig.dev.json",
-    "check:lint": "eslint src *.js",
-    "check:spelling": "cspell \"{README.md,src/**/*.ts}\"",
-    "test": "NODE_ENV=test run-s test:*",
-    "test:node": "NODE_ENV=test TS_NODE_PROJECT=./tsconfig.dev.json mocha",
-    "test:browser": "NODE_ENV=test karma start karma.conf.cjs",
-    "watch:build": "tsc -p tsconfig.json -w",
-    "watch:test": "mocha --watch",
-    "prepublish": "npm run build",
-    "reset-hard": "git clean -dfx -e .idea && git reset --hard && npm i && npm run build"
-  },
-  "engines": {
-    "node": ">=18"
-  },
-  "dependencies": {
-    "@noble/hashes": "^1.3.2",
-    "@waku/utils": "0.0.14"
-  },
-  "devDependencies": {
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-json": "^6.0.0",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@types/chai": "^4.3.11",
-    "@types/debug": "^4.1.12",
-    "@types/mocha": "^10.0.1",
-    "@waku/build-utils": "*",
-    "@waku/interfaces": "0.0.21",
-    "chai": "^4.3.10",
-    "cspell": "^7.3.2",
-    "fast-check": "^3.14.0",
-    "ignore-loader": "^0.1.2",
-    "isomorphic-fetch": "^3.0.0",
-    "mocha": "^10.2.0",
-    "npm-run-all": "^4.1.5",
-    "process": "^0.11.10",
-    "rollup": "^4.9.5"
-  },
-  "files": [
-    "dist",
-    "bundle",
-    "src/*.ts",
-    "src/lib/**/*.ts",
-    "!**/*.spec.*",
-    "!**/*.json",
-    "CHANGELOG.md",
-    "LICENSE",
-    "README.md"
-  ]
-}
